requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
waitress==2.1.2
slack_sdk>=3.0
slack_bolt>=1.6.1
requests==2.31.0
